Add projects/modal-api for the Modal apps and release CLIs
Move the Modal worker/gateway/canary apps, Google credentials setup,
release manifest CLIs, and analytics revision reader into the
policyengine_household_modal project. The country-package install
specs and tax-benefit-system snapshot helpers move to the core
package's new deployment module, which the Cloud Run worker image
imports without the modal SDK.

Image builds switch from Image.uv_sync (which does not support uv
workspaces) to exporting the member's locked closure with uv export
and installing it via uv_pip_install; first-party packages attach as
local Python sources from the deploy machine's all-packages checkout.
The Modal gateway image's dependency list is now lockfile-pinned
instead of hand-maintained.

Part of #1605
