--- libs/household-api/policyengine_household_api/modal_release/__init__.py
@@ -1,1 +1,0 @@
-"""Modal release helpers for current/frontier household API deployments."""
