+++ projects/modal-api/policyengine_household_modal/rewrite_manifest.py
