+++ projects/modal-api/pyproject.toml
@@ -1,0 +1,38 @@
+[project]
+name = "policyengine-household-modal-api"
+version = "0.26.1"
+description = "Modal worker, gateway, and canary apps for the PolicyEngine Household API"
+readme = "README.md"
+authors = [
+    { name = "PolicyEngine", email = "hello@policyengine.org" },
+]
+requires-python = ">=3.12"
+classifiers = ["Private :: Do Not Upload"]
+# Base closure = the Modal gateway image (no country model packages). The
+# worker image installs the `worker` extra, pulling the core application and
+# (transitively) the analytics stack the worker's revision checks use.
+dependencies = [
+    "policyengine-household-common==0.26.1",
+    "flask>=2.2",
+    "modal>=1.3.0",
+    "policyengine-observability[flask]>=1.0.0",
+    "pyyaml>=6",
+]
+
+[project.optional-dependencies]
+worker = [
+    "policyengine-household-api==0.26.1",
+    "policyengine-household-analytics==0.26.1",
+]
+
+[tool.uv.sources]
+policyengine-household-common = { workspace = true }
+policyengine-household-api = { workspace = true }
+policyengine-household-analytics = { workspace = true }
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.wheel]
+packages = ["policyengine_household_modal"]
