+++ tests/unit/modal_release/test_canary_app.py
@@ -2,7 +2,7 @@ import importlib
 
 
 def test_canary_payload_is_static_health_signal():
-    from policyengine_household_api.modal_release import canary_app
+    from policyengine_household_modal import canary_app
 
     assert canary_app.canary_payload() == {
         "ok": True,
@@ -15,7 +15,7 @@ def test_canary_function_options_are_lightweight(monkeypatch):
         "HOUSEHOLD_MODAL_CANARY_APP_NAME",
         "household-canary-test",
     )
-    from policyengine_household_api.modal_release import canary_app
+    from policyengine_household_modal import canary_app
 
     reloaded = importlib.reload(canary_app)
     options = reloaded.canary_function_options()
