+++ tests/unit/modal_release/test_manifest.py
@@ -17,7 +17,7 @@ from policyengine_household_common.release_manifest import (
     rewrite_manifest_for_storage,
     validate_manifest,
 )
-from policyengine_household_api.modal_release import (
+from policyengine_household_modal import (
     manifest as manifest_module,
 )
 from policyengine_household_common.release_config import (
