+++ tests/unit/modal_release/test_rewrite_manifest.py
@@ -3,7 +3,7 @@ import pytest
 from policyengine_household_common.release_manifest import (
     MANIFEST_DICT_KEY,
 )
-from policyengine_household_api.modal_release.rewrite_manifest import (
+from policyengine_household_modal.rewrite_manifest import (
     rewrite_modal_manifest,
 )
